chore: remove old homepage component
